requirements.txt
Flask>=2.0
SQLAlchemy>=1.4
flask-cors>=3.0
requests>=2.28
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
